refactor: Использовать batchUpdate для жанров
